Extract classes from UniqueDataProvider.kt
